+++ db/migrate/20190919202943_create_trips.rb
@@ -4,7 +4,7 @@ class CreateTrips < ActiveRecord::Migration[5.2]
       t.integer :driver_id
       t.integer :user_id
       t.string :time_booked
-      t.integer :total
+      t.float :total
       t.text :review
       t.integer :rating
