+++ db/migrate/20190919202958_create_flat_rates.rb
@@ -2,7 +2,7 @@ class CreateFlatRates < ActiveRecord::Migration[5.2]
   def change
     create_table :flat_rates do |t|
       t.string :route
-      t.integer :duration
+      t.float :duration
 
       t.timestamps
     end
